chore: standardize parameter naming from filePath to fileUri
- Rename filePath parameters to fileUri throughout ApexSymbolManager
- Update method signatures and internal references for consistency
- Remove unused name parameter from resolveAmbiguousSymbolWithContext
- Remove unused generateMigrationPath function (leftover from refactor)
- Maintain backward compatibility with existing interfaces
