+++ packages/apex-parser-ast/src/parser/listeners/BaseApexParserListener.ts
@@ -65,9 +65,9 @@ export abstract class BaseApexParserListener<T> implements ApexParserListener {
   /**
    * Set the current file path for this parser listener.
    * Used for symbol creation and FQN calculation.
-   * @param filePath The file path to set
+   * @param fileUri The file URI to set
    */
-  setCurrentFilePath(filePath: string): void {
+  setCurrentFileUri(fileUri: string): void {
     // This is a no-op in the base class, but subclasses can override
     // to store the file path for symbol creation
   }
