+++ packages/apex-parser-ast/src/symbols/ApexSymbolProcessingManager.ts
@@ -72,22 +72,22 @@ export class ApexSymbolProcessingManager {
    * Store per-file comment associations via the underlying symbol manager.
    */
   setCommentAssociations(
-    filePath: string,
+    fileUri: string,
     associations: import('../parser/listeners/ApexCommentCollectorListener').CommentAssociation[],
   ): void {
-    this.symbolManager.setCommentAssociations(filePath, associations);
+    this.symbolManager.setCommentAssociations(fileUri, associations);
   }
 
   /**
    * Queue a symbol table for background processing
    * @param symbolTable The symbol table to process
-   * @param filePath The file path associated with the symbol table
+   * @param fileUri The file path associated with the symbol table
    * @param options Processing options
    * @returns Task ID for tracking
    */
   processSymbolTable(
     symbolTable: SymbolTable,
-    filePath: string,
+    fileUri: string,
     options: SymbolProcessingOptions = {},
   ): string {
     if (!this.isInitialized) {
@@ -96,17 +96,17 @@ export class ApexSymbolProcessingManager {
           'ApexSymbolProcessingManager not initialized, processing synchronously',
       );
       // Fallback to synchronous processing
-      this.symbolManager.addSymbolTable(symbolTable, filePath);
+      this.symbolManager.addSymbolTable(symbolTable, fileUri);
       return 'sync_fallback';
     }
 
     const taskId = this.symbolIndexingService.processSymbolTable(
       symbolTable,
-      filePath,
+      fileUri,
       options,
     );
     this.logger.debug(
-      () => `Symbol processing queued: ${taskId} for ${filePath}`,
+      () => `Symbol processing queued: ${taskId} for ${fileUri}`,
     );
     return taskId;
   }
@@ -115,13 +115,13 @@ export class ApexSymbolProcessingManager {
    * Schedule persistence of comment associations for a file.
    */
   scheduleCommentAssociations(
-    filePath: string,
+    fileUri: string,
     associations: import('../parser/listeners/ApexCommentCollectorListener').CommentAssociation[],
   ): string {
     // Delegate to the indexing service for background persistence
     // @ts-ignore - access integration
     return this.symbolIndexingService.scheduleCommentAssociations(
-      filePath,
+      fileUri,
       associations,
       'NORMAL',
     );
